requirements_venv.txt
allure-pytest==2.3.2b1
allure-python-commons==2.3.2b1
assertpy==0.13
attrs==18.1.0
backports.functools-lru-cache==1.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
crayons==0.1.2
decorator==4.3.0
enum34==1.1.6
Faker==0.8.15
Flask==1.0.2
future==0.16.0
gevent==1.3.0
greenlet==0.4.13
idna==2.6
ImageHash==4.0
itsdangerous==0.24
Jinja2==2.10
linecache2==1.0.0
locust==0.8
lxml==4.2.1
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.1.0
msgpack-python==0.5.6
namedlist==1.7
numpy==1.14.3
page-objects==1.1.0
pandas==0.23.0
pbr==4.0.3
Pillow==5.1.0
pluggy==0.6.0
py==1.5.3
PyAutoGUI==0.9.36
pycparser==2.18
PyHamcrest==1.9.0
PyMsgBox==1.0.6
PyScreeze==0.1.14
pyshould==0.7.1
pytest==3.5.1
pytest-allure-adaptor==1.7.10
pytest-fixture-config==1.3.0
pytest-webdriver==1.3.0
python-dateutil==2.7.3
PyTweening==1.0.3
pytz==2018.4
PyWavelets==0.5.2
pyzmq==17.0.0
requests==2.18.4
scipy==1.1.0
selene==1.0.0a13
selenium==2.53.1
six==1.11.0
sure==1.4.11
testify==0.11.0
text-unidecode==1.2
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.22
Wand==0.4.4
webdriver-manager==1.6.2
Werkzeug==0.14.1
